+++ src/Ydb.Sdk/CHANGELOG.md
@@ -1,3 +1,6 @@
+- Fixed bug: SQL parser skips token after param. 
+- ADO.NET: Added support for conversion from IN (?, ?, ?) to IN $list ([#447](https://github.com/ydb-platform/ydb-dotnet-sdk/issues/447)).
+
 ## v0.19.0
 
 - ADO.NET: session is now deactivated when cancelled.